+++ app/Http/Controllers/BancaController.php
@@ -30,6 +30,7 @@ class BancaController extends Controller
      */
     public function create($agendamento)
     {
+        $this->authorize('logado');
         $banca = new Banca;
         return view('agendamentos.bancas.create',compact(['agendamento','banca']));
     }
@@ -42,6 +43,7 @@ class BancaController extends Controller
      */
     public function store(BancaRequest $request, Agendamento $agendamento)
     {
+        $this->authorize('logado');
         $banca = new Banca;
         $validated = $request->validated();
         $banca->codpes = $validated['codpes'];
@@ -71,6 +73,7 @@ class BancaController extends Controller
      */
     public function edit(Agendamento $agendamento, Banca $banca)
     {
+        $this->authorize('logado');
         return view('agendamentos.bancas.edit', compact(['agendamento','banca'], [$agendamento,$banca]));
     }
 
@@ -83,6 +86,7 @@ class BancaController extends Controller
      */
     public function update(Agendamento $agendamento, Banca $banca, BancaRequest $request)
     {
+        $this->authorize('logado');
         $validated = $request->validated();
         $banca->agendamento_id = $agendamento->id;
         $banca->update($validated);
@@ -97,6 +101,7 @@ class BancaController extends Controller
      */
     public function destroy(Agendamento $agendamento, Banca $banca)
     {
+        $this->authorize('logado');
         $banca->delete();
         return redirect("/agendamentos/{$agendamento->id}");
     }
